# backend-main/app/openapi.yml
openapi: '3.0.2'
info:
  title: SAM API spec
  version: '1.0'
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []  
paths:
...
  /products:
    get:
      tags:
        - Payment
      summary: Get all products
      description: Get all products
      security: []
      responses:
        200:
          description: products succesfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: "stripe_product_id"
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"
  /price/{id}:
    get:
      tags:
        - Payment
      summary: Get price by id
      description: Get price by id
      security: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
            example: "stripe_price_id"
          required: true
          description: The id of the price
      responses:
        200:
          description: price succesfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "stripe_price_id"
                  active:
                    type: boolean
                    example: true
                  product:
                    type: string
                    example: "stripe_product_id"
                  unit_amount:
                    type: number
                    example: 1000
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"
  /customer_portal:
    get:
      tags:
        - Payment
      summary: Get stripe customer portal
      description: Get stripe customer portal
      responses:
        302:
          description: Redirect to stripe customer portal url
          headers:
            Location:
              description: The URL to redirect to
              schema:
                type: string
        500:
          description: Internal Server Error
          content:
            plain/text:
              schema:
                type: string
                example: "Internal Server Error"
  /create_payment:
    post:
      tags:
        - Payment
      summary: Create a payment
      description: Create a payment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "stripe_item_id"
                      quantity:
                        type: number
                        format: int32
                        example: 1
      responses:
        200:
          description: Redirect to stripe checkout url
          content:
            plain/text:
              schema:
                type: string
                example: "url"
        400:
          description: Can't create payment
          content:
            plain/text:
              schema:
                $ref: '#/components/schemas/badRequest'
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT 
  schemas:
...
    badRequest:
      type: string
      example: "Bad Request"
...